+++ tests/integration_tests/mod.rs
@@ -3,4 +3,5 @@
 #[path = "../common/mod.rs"]
 mod common;
 
+mod mcp_test;
 mod server_test;